+++ concrete_datastore/admin/admin.py
@@ -221,6 +221,7 @@ class EmailDeviceAdmin(SaveModelMixin, admin.ModelAdmin):
         'uid',
         'user',
         'key',
+        'confirmed',
         'mfa_mode',
         'digits',
         'step',