+++ concrete_datastore/__init__.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 
 
-VERSION = (1, 51, 0)
+VERSION = (1, 52, 0)
 
 
 def get_version(version=VERSION):