+++ CHANGELOG.md
@@ -8,13 +8,19 @@
 
 ### Changed
 
-- Management command to invite a user to register a TOTP device now supports a list of emails
-- Setting to change email subject for otp activation
-- Use Email MFA in order to activate OTP MFA to add more security
+- nothing changed
 
 ### Removed
 
 - nothing removed
+
+## [1.52.0] - 2022-12-12
+
+### Changed
+
+- Management command to invite a user to register a TOTP device now supports a list of emails
+- Setting to change email subject for otp activation
+- Use Email MFA in order to activate OTP MFA to add more security
 
 ## [1.51.0] - 2022-11-25